Basic knockback implemented
KnockBack() created in Health.cs script.
This function finds the vector between the player and the given damage source, normalizes it, and applies a given magnitude to control the distance which the player is knocked back.

The version might have changed because of the school's version of unity, but it shouldn't cause any problems.