Raycast Detection implemented, shooting in progress
Raycast detection is implemented. To make it work, all you need to do is mark the tilemaps that shouldn't be seen through as "obstacles" in the layers drop down in the inspector.

Renamed my scene to kasey-old.

Negligable changes to the playtest scene. Apologies in advance if this screws something up, the changes are minimal.