Shooting implemented
To use the shooting script, take an enemy object or prefab, add the component EnemyShooting, then select the prefab that you want to be the projectile (currently EnemyProjectile is the only one that works), and the shooting will work. You can adjust the offset for the shot origin point and the rate of fire. The speed of the projectile is handled in the given projectile prefab's script. Currently the EnemyProjectile prefab is set to 20 speed. The Kasey scene has an implementation of the shooting script ready to use.