Pathfinding implementation, Shooting correction
Astar pathfinding has been implemented in a mostly working fashion. Prefabs have been updated to be compatible, however be watchful of bugs.

Note for grading purposes: Most of the files here are NOT mine, it's a publicly available pathfinding system for Unity called Astar. The implementation in the game's script files are my work.

Shooting has also been corrected to aim at the player's proper position.